FIX: Use double to represent a radius in collisions.

+++ game/unit/damageable.h
@@ -12,7 +12,7 @@ public:
 	virtual auto collide(const segment_t&) -> std::optional<std::pair<
 		std::reference_wrapper<interface::damageable>,
 		vector<double>>> = 0;
-	virtual auto collide(vector<double> center, float radius)
+	virtual auto collide(vector<double> center, double radius)
 		-> std::vector<std::reference_wrapper<interface::damageable>> = 0;
 	virtual auto radius() const -> double = 0;
